Split the util module into smaller parts
